Fixed and finalized details

+++ src/CloudFaceDetection/CloudFaceScripts/aMatching.cs
@@ -4,5 +4,5 @@
 public class aMatching
 {
     public bool isIdentical;
-    public float confidence;
+    public double confidence;
 }
